+++ src/languageServerPlugin.ts
@@ -3,10 +3,10 @@
 import * as commands from "./commands";
 
 export enum CompileWorkspaceStatus {
-    FAILED = 0,
-    SUCCEED = 1,
-    WITHERROR = 2,
-    CANCELLED = 3,
+    Failed = 0,
+    Succeed = 1,
+    Witherror = 2,
+    Cancelled = 3,
 }
 
 export function resolveBuildFiles(): Promise<string[]> {
